+++ src/Custom/Assistants/Streaming/RunStepDetailsUpdateCodeInterpreterOutput.cs
@@ -6,7 +6,7 @@ namespace OpenAI.Assistants;
 public partial class RunStepUpdateCodeInterpreterOutput
 {
     /// <inheritdoc cref="InternalRunStepDeltaStepDetailsToolCallsCodeOutputLogsObject.Index"/>
-    public int OutputIndex => AsLogs?.Index ?? AsImage?.Index ?? 0;
+    public long OutputIndex => AsLogs?.Index ?? AsImage?.Index ?? 0;
 
     /// <inheritdoc cref="InternalRunStepDeltaStepDetailsToolCallsCodeOutputLogsObject.InternalLogs"/>
     public string Logs => AsLogs?.InternalLogs;
