+++ src/Custom/Chat/ChatMessageAnnotation.cs
@@ -7,9 +7,9 @@ namespace OpenAI.Chat;
 [CodeGenType("ChatCompletionResponseMessageAnnotation")]
 public partial class ChatMessageAnnotation
 {
-    public int StartIndex => UrlCitation?.StartIndex
+    public long StartIndex => UrlCitation?.StartIndex
         ?? 0;
-    public int EndIndex => UrlCitation?.EndIndex
+    public long EndIndex => UrlCitation?.EndIndex
         ?? 0;
     public Uri WebResourceUri => UrlCitation?.Url;
     public string WebResourceTitle => UrlCitation?.Title;
