+++ src/Custom/Responses/OpenAIResponseClient.Protocol.cs
@@ -6,15 +6,15 @@ using System.Threading.Tasks;
 
 namespace OpenAI.Responses;
 
-[CodeGenSuppress("GetResponse", typeof(string), typeof(string), typeof(IEnumerable<InternalIncludable>), typeof(bool?), typeof(int?), typeof(RequestOptions))]
-[CodeGenSuppress("GetResponseAsync", typeof(string), typeof(string), typeof(IEnumerable<InternalIncludable>), typeof(bool?), typeof(int?), typeof(RequestOptions))]
-[CodeGenSuppress("CancelResponse", typeof(string), typeof(IEnumerable<InternalIncludable>), typeof(bool?), typeof(int?), typeof(RequestOptions))]
-[CodeGenSuppress("CancelResponseAsync", typeof(string), typeof(IEnumerable<InternalIncludable>), typeof(bool?), typeof(int?), typeof(RequestOptions))]
-[CodeGenSuppress("GetResponse", typeof(string), typeof(IEnumerable<InternalIncludable>), typeof(bool?), typeof(int?), typeof(RequestOptions))]
-[CodeGenSuppress("GetResponseAsync", typeof(string), typeof(IEnumerable<InternalIncludable>), typeof(bool?), typeof(int?), typeof(RequestOptions))]
+[CodeGenSuppress("GetResponse", typeof(string), typeof(string), typeof(IEnumerable<InternalIncludable>), typeof(bool?), typeof(long?), typeof(RequestOptions))]
+[CodeGenSuppress("GetResponseAsync", typeof(string), typeof(string), typeof(IEnumerable<InternalIncludable>), typeof(bool?), typeof(long?), typeof(RequestOptions))]
+[CodeGenSuppress("CancelResponse", typeof(string), typeof(IEnumerable<InternalIncludable>), typeof(bool?), typeof(long?), typeof(RequestOptions))]
+[CodeGenSuppress("CancelResponseAsync", typeof(string), typeof(IEnumerable<InternalIncludable>), typeof(bool?), typeof(long?), typeof(RequestOptions))]
+[CodeGenSuppress("GetResponse", typeof(string), typeof(IEnumerable<InternalIncludable>), typeof(bool?), typeof(long?), typeof(RequestOptions))]
+[CodeGenSuppress("GetResponseAsync", typeof(string), typeof(IEnumerable<InternalIncludable>), typeof(bool?), typeof(long?), typeof(RequestOptions))]
 public partial class OpenAIResponseClient
 {
-    public virtual async Task<ClientResult> GetResponseAsync(string responseId, bool? stream, int? startingAfter, RequestOptions options)
+    public virtual async Task<ClientResult> GetResponseAsync(string responseId, bool? stream, long? startingAfter, RequestOptions options)
     {
         Argument.AssertNotNullOrEmpty(responseId, nameof(responseId));
 
@@ -35,7 +35,7 @@ public partial class OpenAIResponseClient
 
     public virtual AsyncCollectionResult GetResponseInputItemsAsync(
         string responseId,
-        int? limit,
+        long? limit,
         string order,
         string after,
         string before,
@@ -48,7 +48,7 @@ public partial class OpenAIResponseClient
 
     public virtual CollectionResult GetResponseInputItems(
         string responseId,
-        int? limit,
+        long? limit,
         string order,
         string after,
         string before,
