+++ src/Custom/VectorStores/VectorStoreExpirationPolicy.cs
@@ -8,8 +8,8 @@ namespace OpenAI.VectorStores;
 /// Represents the the configuration that controls when a vector store will be automatically deleted.
 /// </summary>
 [CodeGenType("VectorStoreExpirationAfter")]
-[CodeGenSuppress(nameof(VectorStoreExpirationPolicy), typeof(int))]
-[CodeGenSuppress(nameof(VectorStoreExpirationPolicy), typeof(int), typeof(VectorStoreExpirationAnchor))]
+[CodeGenSuppress(nameof(VectorStoreExpirationPolicy), typeof(long))]
+[CodeGenSuppress(nameof(VectorStoreExpirationPolicy), typeof(long), typeof(VectorStoreExpirationAnchor))]
 public partial class VectorStoreExpirationPolicy
 {
     [CodeGenMember("Anchor")]
